+++ Assets/Content/Features/MainScreenModule/Scripts/ISudokuGridUserIntentListener.cs
@@ -4,7 +4,7 @@ namespace Content.Features.MainScreenModule.Scripts
 {
     public interface ISudokuGridUserIntentListener
     {
-        public event Action<int> OnGridEditRequested;
+        public event Action<long> OnGridEditRequested;
         public event Action OnGridCreationRequested;
     }
 }
